+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased]
 
+### Changed
+
+- The `propertyNames` validator now contains the parent object in its `instance` attribute instead of individual properties as strings.
+- Improved error message for the `additionalProperties` validator. After - `Additional properties are not allowed ('faz' was unexpected)`, before - `False schema does not allow '"faz"'`.
+- The `additionalProperties` validator emits a single error for all unexpected properties instead of separate errors for each unexpected property.
+
+### Fixed
+
+- All `instance_path` attributes are pointing to the proper location.
+
 ## [0.7.0] - 2021-04-27
 
 ### Added